tests: Renamed to encompass SchemaConfiguration and CacheKeyProvider
